bots/praw_without_venv/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.9
pathspec==0.7.0
pkg-resources==0.0.0
praw==6.5.1
prawcore==1.0.1
python-dotenv==0.12.0
regex==2020.2.20
requests==2.23.0
six==1.14.0
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
